exclusões e alterações de layout

+++ src/main/java/br/com/panoramico/managebean/cadastro/AssociadoMB.java
@@ -7,8 +7,10 @@ package br.com.panoramico.managebean.cadastro;
 
 import br.com.panoramico.dao.AssociadoDao;
 import br.com.panoramico.dao.AssociadoEmpresaDao;
+import br.com.panoramico.dao.DependenteDao;
 import br.com.panoramico.model.Associado;
 import br.com.panoramico.model.Associadoempresa;
+import br.com.panoramico.model.Dependente;
 import br.com.panoramico.uil.Mensagem;
 import java.io.Serializable;
 import java.sql.SQLException;
@@ -43,7 +45,8 @@ public class AssociadoMB implements Serializable{
     @EJB
     private AssociadoEmpresaDao associadoEmpresaDao;
     private Associadoempresa associadoempresa;
-    
+    @EJB
+    private DependenteDao dependenteDao;
     
     @PostConstruct
     public void init(){
@@ -88,6 +91,14 @@ public class AssociadoMB implements Serializable{
 
     public void setAssociadoempresa(Associadoempresa associadoempresa) {
         this.associadoempresa = associadoempresa;
+    }
+
+    public DependenteDao getDependenteDao() {
+        return dependenteDao;
+    }
+
+    public void setDependenteDao(DependenteDao dependenteDao) {
+        this.dependenteDao = dependenteDao;
     }
     
     
@@ -157,8 +168,13 @@ public class AssociadoMB implements Serializable{
     }
     
     public void excluir(Associado associado){
-        associadoDao.remove(associado.getIdassociado());
-        Mensagem.lancarMensagemInfo("Excluido", "com sucesso");
-        gerarListaAssociado();
+        List<Dependente> listaDependente = dependenteDao.list("Select d from Dependente d where d.associado.idassociado="+ associado.getIdassociado());
+        if (listaDependente == null || listaDependente.isEmpty()) {
+            associadoDao.remove(associado.getIdassociado());
+            Mensagem.lancarMensagemInfo("Excluido", "com sucesso");
+            gerarListaAssociado();
+        }else{
+            Mensagem.lancarMensagemInfo("Atenção", "este associado não pode ser excluido");
+        }
     }
 }
